pyproject.toml
[tool.poetry]
name = "99shops"
version = "0.1.0"
description = ""
authors = ["prasanta-mohapatra <mohapatraprasant98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.97.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
taskipy = "^1.11.0"
sqlalchemy = "^1.4.41"
pytest = "^7.3.2"
python-dotenv = "^1.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
alembic = "^1.11.1"
psycopg2 = "^2.9.6"
sqlmodel = "^0.0.8"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
app = "uvicorn main:app --host 0.0.0.0 --port 8000 --reload"
